misc: drv8846: reject popup movement requests during bootup
Somehow when the device is booting selinux enforced, the xiaomi motor
service is sending the kernel driver the command to move up several
times. Logging does not show any information why this is happening.
When Booting selinux permissive, this is not happening.

Until the real reason is found, this dirty hack is used to be able to
boot selinux enforced without the camera popping out during boot.

[yaro: improved code quality a bit, shortened commit message]
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>